Code Cleanup & Modularization
- Clean up and Modularization of Popup
- Modularized Content [TODO: Clean up code]
1. TODO: Clean and Modularize Background Script
