+++ Scripts/PopupUI/modules/messaging.js
@@ -1,0 +1,23 @@
+export function sendContentMessage(activeTabId, message) {
+    return new Promise((resolve, reject) => { 
+        chrome.tabs.sendMessage(activeTabId, message, (response) => {
+            if (chrome.runtime.lastError) {
+                reject(chrome.runtime.lastError);
+            } else {
+                resolve(response);
+            }
+        });
+    });
+}
+
+export function sendRuntimeMessage(message) {
+    return new Promise((resolve, reject) => {
+        chrome.runtime.sendMessage(message, (response) => {
+            if (chrome.runtime.lastError) {
+                reject(chrome.runtime.lastError);
+            } else {
+                resolve(response);
+            }
+        });
+    });
+}
