+++ Scripts/PopupUI/popup.html
@@ -62,6 +62,6 @@
         </div>
     </div>
 
-    <script src="popup.js"></script>
+    <script type="module" src="popup.js"></script>
 </body>
 </html>
